fmpq_poly set
